+++ ledger-service/http-json-testing/BUILD.bazel
@@ -30,11 +30,7 @@ hj_scalacopts = lf_scalacopts + [
             "@maven//:org_scalacheck_scalacheck",
             "@maven//:org_scalactic_scalactic",
             "@maven//:org_scalatest_scalatest",
-            "@maven//:org_tpolecat_doobie_core",
-            "@maven//:org_tpolecat_doobie_free",
-            "@maven//:org_typelevel_cats_core",
             "@maven//:org_typelevel_cats_effect",
-            "@maven//:org_typelevel_cats_free",
         ],
         scalacopts = hj_scalacopts,
         tags = ["maven_coordinates=com.daml:http-json-testing:__VERSION__"],
