+++ ledger-service/http-json/BUILD.bazel
@@ -253,8 +253,11 @@ alias(
             "@maven//:org_scalaz_scalaz_core",
             "@maven//:org_tpolecat_doobie_core",
             "@maven//:org_tpolecat_doobie_hikari",
+            "@maven//:org_tpolecat_doobie_free",
             "@maven//:org_typelevel_cats_core",
             "@maven//:org_typelevel_cats_effect",
+            "@maven//:org_typelevel_cats_free",
+            "@maven//:org_typelevel_cats_kernel",
         ],
         visibility = ["//ledger-service:__subpackages__"],
         deps = [
@@ -268,6 +271,7 @@ alias(
             "//ledger-api/rs-grpc-bridge",
             "//ledger-service/http-json-cli:{}".format(edition),
             "//ledger-service/http-json-testing:{}".format(edition),
+            "//ledger-service/db-backend",
             "//ledger-service/jwt",
             "//ledger-service/utils",
             "//ledger/ledger-api-auth",
